backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.6.0
django-filter==23.4
djangorestframework-simplejwt==5.3.0
